Move Purchase and Play logic into custom hook
This will make it easier to reuse in other post-checkout contexts (e.g.
the inline checkout for a lesson). It also helps isolate some
dependencies.
